backend/requirements.txt
fastapi
uvicorn
ollama
requests
pydantic
python-dotenv
python-multipart
pdfplumber
python-docx
openpyxl
